Add condition via javascript.
This change is to be able to add an additional field to be taken into account as a condition when creating a "parent select".
description:
I have a select field called "color1", a select field named "color2" and a third field named flag (pointing to a flag table with their respective colors in different fields). In this third field I can assign the values "color1" and "color2" to the key "parent_select" to the array "form". Which will return the flags that match these 2 colors.
I.e:

´$colors = ['yellow','blue',...];´

´$this->form[] = ['label'=>'Color 1','name'=>'color_1','type'=>'select','validation'=>'required|integer|min:0','width'=>'col-sm-10','dataenum'=>$colors];´

´$this->form[] = ['label'=>'Color 2','name'=>'color_2','type'=>'select','validation'=>'required|integer|min:0','width'=>'col-sm-10','datatable'=>$colors];´

´$this->form[] = ['label'=>'Flag','name'=>'id_flags','type'=>'select','validation'=>'required|integer|min:0','width'=>'col-sm-10','datatable'=>'flags,flag','parent_select'=>'color_1,color_2'];´