Update CBController, Support Post Values
The insert column counts become offset in cases when you have zero values. Using "empty" failed in conditions when the field is not required. The insert query builder was set to insert the array, which must match column count.  Using "isset" prompts the business logic to gather the value regardless.